% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assists_fibalivestats.R
\name{get_assists_fibalivestats}
\alias{get_assists_fibalivestats}
\title{Download assists from fibalivestats.com using game id}
\usage{
get_assists_fibalivestats(gameid)
}
\arguments{
\item{gameid}{numeric, game id from fibalivestats.com}
}
\value{
data.frame with assists
}
\description{
Download assists from fibalivestats.com using game id
}
\examples{
get_assists_fibalivestats(742430)

}
\seealso{
\code{\link[euRobasket]{get_assists_livefibaeurope}}, @seealso \code{\link[euRobasket]{assist_network}}
}
