% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assists_livefibaeurope.R
\name{get_assists_livefibaeurope}
\alias{get_assists_livefibaeurope}
\title{Download assists from live.fibaeurope.com using game id}
\usage{
get_assists_livefibaeurope(gameid)
}
\arguments{
\item{gameid}{numeric, game id from live.fibaeurope.com}
}
\value{
data.frame with assists
}
\description{
Download assists from live.fibaeurope.com using game id
}
\examples{
get_assists_livefibaeurope(742430)

}
\seealso{
\code{\link[euRobasket]{get_assists_fibalivestats}}, @seealso \code{\link[euRobasket]{assist_network}}
}
