% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_box_score.R
\name{get_boxscores_realgm}
\alias{get_boxscores_realgm}
\title{Download boxscores from realgm.com for given period of time}
\usage{
get_boxscores_realgm(start_date, end_date, league)
}
\arguments{
\item{start_date}{Date in format YYYY-MM-DD}

\item{end_date}{Date in format YYYY-MM-DD}

\item{league}{name of league, all names can be found using function
get_realgm_league_names()}
}
\value{
list of box scores
}
\description{
Download boxscores from realgm.com for given period of time
}
\examples{

#Load box scores from Spanish ACB
get_boxscores_realgm('2017-10-01', '2017-10-08', 'Spanish ACB')

#Load box scores from Austrian League
#Find exact name of the league using get_realgm_league_names()
get_realgm_league_names()
get_boxscores_realgm('2016-01-12', '2016-09-11', 'Austrian A Bundesliga')
}
