% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_on_off_splits_players.R
\name{get_on_off_splits_players}
\alias{get_on_off_splits_players}
\title{Get available players for on_off_splits() from stints data frame}
\usage{
get_on_off_splits_players(stints.df, print = TRUE)
}
\arguments{
\item{stints.df}{data.frame, stints downloaded with get_stints_fibalivestats() or get_stints_livefibaeurope()}

\item{print}{logical, TRUE by default, set to FALSE if you do not want to print results.
For on_off_splits() puropses mostly.}
}
\value{
data.frame with all players and teams available in stints.df
}
\description{
Use this function if you want to find players for on_off_splits.
}
\examples{
#get stints from Basketball Champions League using data()
#it will be stored in cl_stints
data("cl_stints")

#get all different players and teams from this data
get_on_off_splits_players(cl_stints)
}
\seealso{
\code{\link[euRobasket]{get_on_off_splits_players}}, \code{\link[euRobasket]{get_stints_fibalivestats}}, \code{\link[euRobasket]{get_stints_livefibaeurope}}
}
