% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_players_stats_realgm.R
\name{get_players_stats_realgm}
\alias{get_players_stats_realgm}
\title{Load players stats from realgm.com}
\usage{
get_players_stats_realgm(league, season, type, position = "All")
}
\arguments{
\item{league}{character, league name from realgm.com, you can find all
available names using get_realgm_league_names()}

\item{season}{numeric, second year of season, ie. for season 2016-17 use 2017,
for season 2013-14 use 2014}

\item{type}{character, type of stats to downlod, choose from:
Totals, Averages, Per_48, Per_40, Per_36, Per_Minute,
Minute_Per, Misc_Stats, Advanced_Stats}

\item{position}{character, 'All' set by default, change to specific position ie.
if you want to download stats only for point guards use type = 'PG', available positions:
All, PG, SG, SF, PF, C}
}
\value{
data frame with players stats
}
\description{
Load different types of players stats for choosen league and season
from realgm.com. You need to specify league name, season and stats type.
}
\examples{

#Load Advanced Stats for PGs from Spanish ACB, season 2014-15
get_players_stats_realgm('Spanish ACB', 2015, type = 'Advanced_Stats', position = 'PG')

#Load Totals for all players from Polish TBL, season 2012-13
get_players_stats_realgm('Polish TBL', 2013, type = 'Totals')
}
