% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_pbp_fibalivestats.R
\name{get_raw_pbp_fibalivestats}
\alias{get_raw_pbp_fibalivestats}
\title{Download play by play from fibalivestats.com using gameid}
\usage{
get_raw_pbp_fibalivestats(gameid)
}
\arguments{
\item{gameid}{numeric, game id from fibalivestats.com}
}
\value{
data.frame with play by play data
}
\description{
Download play by play from fibalivestats.com using gameid
}
\examples{
#Download play by play for game 412182
get_raw_pbp_fibalivestats(412182)

}
\seealso{
\code{\link[euRobasket]{get_stints_fibalivestats}} \code{\link[euRobasket]{get_raw_pbp_livefibaeurope}}
}
