% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_pbp_livefibaeurope.R
\name{get_raw_pbp_livefibaeurope}
\alias{get_raw_pbp_livefibaeurope}
\title{Download play by play from live.fibaeurope.com using gameid}
\usage{
get_raw_pbp_livefibaeurope(gameid)
}
\arguments{
\item{gameid}{numeric, game id from live.fibaeurope.com}
}
\value{
data.frame with play by play data
}
\description{
Download play by play from live.fibaeurope.com using gameid
}
\examples{
#Download play by play for game 412182
get_raw_pbp_livefibaeurope(412182)
}
\seealso{
\code{\link[euRobasket]{get_raw_pbp_fibalivestats}}\code{\link[euRobasket]{get_stints_livefibaeurope}}
}
