% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scores_realgm.R
\name{get_scores_realgm}
\alias{get_scores_realgm}
\title{Download scores from realgm.com for given period of time.
You need to specify start date, end date and league name}
\usage{
get_scores_realgm(start_date, end_date, league)
}
\arguments{
\item{start_date}{Date in format YYYY-MM-DD}

\item{end_date}{Date in format YYYY-MM-DD}

\item{league}{character, league name from realgm.com, you can find all
available names using get_realgm_league_names()}
}
\value{
data frame with scores
}
\description{
Download scores from realgm.com for given period of time.
You need to specify start date, end date and league name
}
\examples{

#Load scores from Israeli BSL, first week of march 2017
get_scores_realgm('2017-03-01', '2017-03-07', 'Israeli BSL')

}
