% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shots_data_fiba.R
\name{get_shots_data_fibalivestats}
\alias{get_shots_data_fibalivestats}
\title{Load shots data from fibalivestats.com with gameid}
\usage{
get_shots_data_fibalivestats(gameid)
}
\arguments{
\item{gameid}{game ID from fibalivestats.com, numeric}
}
\value{
data frame with shots
}
\description{
Load shots data from fibalivestats.com with gameid
}
\examples{
#Load shots from game 9863
get_shots_data_fibalivestats(9863)

}
\seealso{
\code{\link[euRobasket]{get_shots_data_livefibaeurope}}
}
