% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shots_data_livefibaeurope.R
\name{get_shots_data_livefibaeurope}
\alias{get_shots_data_livefibaeurope}
\title{Load shots data from live.fibaeurope.com with gameid}
\usage{
get_shots_data_livefibaeurope(gameid)
}
\arguments{
\item{gameid}{game ID from live.fibaeurope.com, numeric}
}
\value{
data frame with shots
}
\description{
Load shots data from live.fibaeurope.com with gameid
}
\examples{
#Load shots from game 742430
get_shots_data_livefibaeurope(742430)

}
\seealso{
\code{\link[euRobasket]{get_shots_data_fibalivestats}}
}
