% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stints_fibalivestats.R
\name{get_stints_fibalivestats}
\alias{get_stints_fibalivestats}
\title{Get stints from fibalivestats.com using game id}
\usage{
get_stints_fibalivestats(gameid)
}
\arguments{
\item{gameid}{numeric, game id from fibalivestats.com}
}
\value{
data.frame with stints
}
\description{
Get stints from play by play.
Stint is a period of time that the same set of ten players are on the court.
For every stint this function calculates stats such as total assists, total points, total possessions.
}
\examples{
get_stints_fibalivestats(411996)

#You can also download multiple stints with for loop
#First get ids for the league that you want
In this example it is: FIBA Europe Cup

Get data.frame with ids and league names
data("fibalivestats_matches_ids")

#Find ids for FIBA Europe Cup
ids = fibalivestats_matches_ids$id[fibalivestats_matches_ids$league == 'FIBA Europe Cup']

#get stints with for loop
stints = data.frame()
for(i in 1:length(ids)) {
stints = rbind(stints, get_stints_fibalivestats(ids[i]))
}

}
\seealso{
\code{\link[euRobasket]{get_stints_livefibaeurope}} \code{\link[euRobasket]{on_off_splits}}, \code{\link[euRobasket]{get_on_off_splits_players}}
}
