% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stints_livefibaeurope.R
\name{get_stints_livefibaeurope}
\alias{get_stints_livefibaeurope}
\title{Get stints from live.fibaeurope.com using game id}
\usage{
get_stints_livefibaeurope(gameid)
}
\arguments{
\item{gameid}{numeric, game id from live.fibaeurope.com}
}
\value{
data.frame with stints
}
\description{
Get stints from play by play.
Stint is a period of time that the same set of ten players are on the court.
For every stint this function calculates stats such as total assists, total points, total possessions.
}
\examples{
#Get stints for id 108510
get_stints_livefibaeurope(108510)
}
\seealso{
\code{\link[euRobasket]{get_stints_fibalivestats}} \code{\link[euRobasket]{on_off_splits}}, \code{\link[euRobasket]{get_on_off_splits_players}}
}
