% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineups_stats.R
\name{lineups_stats}
\alias{lineups_stats}
\title{Get stats for lineups from stints}
\usage{
lineups_stats(stints.df, min.possessions = 1, ratings.only = FALSE)
}
\arguments{
\item{stints.df}{data.frame, stints downloaded with get_stints_fibalivestats() or get_stints_livefibaeurope()}

\item{min.possessions}{numeric, set to 1 by default, minimum number of possessions for
lineups to be included in function results}

\item{ratings.only}{logical, set to FALSE by default, if set to TRUE for
each lineup function results will only show number of possessions, offensive rating,
defensive rating and net rating}
}
\value{
data.frame with stats for each lineup
}
\description{
This function calculate stats for each lineup in stints.
Stints can be downloaded using get_stints_fibalivestats() or get_stints_livefibaeurope().
}
\examples{
#get lineups stats for basketball champions league
#firstly get the stints from bcl using data()
data("cl_stints")

#use lineup_stats() to get stats for each lineup that played more than 10 possessions

lineups_stats(cl_stints, min.possessions = 10)
}
\seealso{
\code{\link[euRobasket]{get_stints_fibalivestats}}, \code{\link[euRobasket]{get_stints_livefibaeurope}}
}
