% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_off_splits.R
\name{on_off_splits}
\alias{on_off_splits}
\title{Get on/off splits for selected players from stints.}
\usage{
on_off_splits(stints.df, players)
}
\arguments{
\item{stints.df}{data.frame, stints download with get_stints_fibalivestats() or get_stints_livefibaeurope()}

\item{players}{a character vector with player or players for on/off splits,
Max number of players is 5.}
}
\value{
list with data.frames, data frames contains different stats types:
Ratings, Team Shooting, Opponent Shooting, Team Playmaking, Opponent Playmaking,
Team Boxscore Defensive, Opponent Boxscore Defensive
}
\description{
On/off splits shows how team and opponents played with and without selected players.
For this function selected players needs to be from the same team. You can select
up to 5 players. To get all available players in stints use get_on_off_splits_players()
}
\examples{
#get on/off splits from Basketball Champions League
#firstly get the stints from bcl using data(), stored in cl_stints
data("cl_stints")

#check available players
get_on_off_splits_players(cl_stints)

#get on/off splits for M. PONITKA
on_off_splits(cl_stints, players = 'M. PONITKA')

#get on/off splits for M. PONITKA, D. SUMMERS and J. BROWN
on_off_splits(cl_stints, players = c('M. PONITKA', 'D. SUMMERS', 'J. BROWN'))

}
\seealso{
\code{\link[euRobasket]{get_on_off_splits_players}}, \code{\link[euRobasket]{get_stints_fibalivestats}}, \code{\link[euRobasket]{get_stints_livefibaeurope}}
}
