% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shot_chart.R
\name{shot_chart}
\alias{shot_chart}
\title{Visualize shots data}
\usage{
shot_chart(shots_df, title = "Shot chart", type = "Scatter")
}
\arguments{
\item{shots_df}{data.frame, shots data downloaded with get_shots_data_fibalivestats() or get_shots_data_livefibaeurope()}

\item{title}{character, shot chart title, 'Shot chart' set by default}

\item{type}{character, type of shot chart to create, use 'Scatter' for scatter plot
or 'Heatmap' for heatmap}
}
\description{
Create scatter charts or heatmaps from shots data download with get_shots_data_fibalivestats()
or get_shots_data_livefibaeurope(). This function creates one shot chart from shots data.
If you want to get multiple shot charts use shots_dashboard()
}
\examples{
#get heatmap for M. Ponitka from Basketball Champions League shots example
#get data using data(), it will be stored in cl_shots
data("cl_shots")

#subset M. PONITKA shots
shots = cl_shots[cl_shots$player == 'M. PONITKA',]

#plot heatmap
shot_chart(shots, title = 'M. Ponitka', type = 'Heatmap')
}
\seealso{
\code{\link[euRobasket]{shots_dashboard}}
}
