% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shots_app.R
\name{shots_dashboard}
\alias{shots_dashboard}
\title{Create dashboard to visualize shots}
\usage{
shots_dashboard(shots_df, dashboardTitle = "Shots", save = FALSE,
  app_name = "shots_dashboard")
}
\arguments{
\item{shots_df}{Shots data in data.frame format. It can be downloaded using get_shots_data_livefibaeurope()
or get_shots_data_fibalivestats()}

\item{dashboardTitle}{character, title of dashboard, 'Shots' set by default}

\item{save}{logical, FALSE set by default, use save = TRUE if you want to save app in
new folder in working directory. It can be then posted on shinyapps.io}

\item{app_name}{character, 'shots_dashboard' set by default,
name of the directory where the application will be saved}
}
\description{
This function creates dashboard from shots data loaded with get_shots_data_livefibaeurope()
or get_shots_data_fibalivestats(). It containts two type of shot charts: scatter and heatmap.
It only runs locally by default, but you can use save = TRUE to write app to directory and then
post it on shinyapps.io.
}
\examples{
#create shots dashboard for Basketball Champions League shots example
#get data using data(), it will be stored in cl_shots
data("cl_shots")

#create dashboard, run locally
shots_dashboard(shots_df = cl_shots, dashboardTitle = 'BCL Shots')

#create dashboard and save it in working directory
shots_dashboard(shots_df = cl_shots, dashboardTitle = 'BCL Shots', save = TRUE, app_name = 'BCL_Dashboard')
}
